Move all JNI infra in node-common/java and leave */jni for exports
