+++ core-rust/node-common/src/java/utils.rs
