+++ core-rust/state-manager/src/mempool/mempool_relay_dispatcher.rs
@@ -68,7 +68,7 @@ use jni::{JNIEnv, JavaVM};
 use radix_engine_common::data::scrypto::scrypto_encode;
 use std::ops::Deref;
 
-use node_common::jni::utils::jni_slice_to_jbytearray;
+use node_common::java::*;
 use transaction::model::NotarizedTransaction;
 
 use crate::jni::mempool::JavaRawTransaction;
